jobs models and routes completed

+++ models/Job.js
@@ -5,7 +5,7 @@ const APIError = require('../helpers/apiError');
 class Job {
   // model for enacting crud on jobs
 
-  static async getJobs({ search, min_salary = 0, max_salary = Infinity }) {
+  static async getJobs({ search, min_salary = 0, max_salary = 9999999999999 }) {
     // Titles and company handles for all jobs, ordered by most recent
     /* Query string params: 
           search: filtered list of titles and company handles
@@ -20,7 +20,7 @@ class Job {
         WHERE salary BETWEEN $2 AND $3
         AND lower(title) LIKE lower($1)
         ORDER BY date_posted DESC`,
-        [search, min_salary, max_salary]
+        [`%${search}%`, min_salary, max_salary]
       );
 
       if (result.rows.length === 0) {
@@ -67,6 +67,10 @@ class Job {
 
   // GET job by id
   static async getJobById(id) {
+    if (typeof id !== 'number') {
+      throw new APIError('Invalid id for job', 404);
+    }
+
     const result = await db.query(
       `SELECT * FROM jobs
       WHERE id = $1`,
@@ -74,7 +78,7 @@ class Job {
     );
 
     if (result.rows.length === 0) {
-      throw new Error('No such company exists');
+      throw new APIError('No such company exists', 404);
     }
 
     return result.rows[0];
@@ -82,6 +86,10 @@ class Job {
 
   // PATCH job by id
   static async updateJobById(id, data) {
+    if (typeof id !== 'number') {
+      throw new APIError('Invalid id for job', 404);
+    }
+
     const { query, values } = sqlForPartialUpdate(
       'jobs',
       data.values,
@@ -92,7 +100,7 @@ class Job {
     const result = await db.query(query, values);
 
     if (result.rows.length === 0) {
-      throw new Error('No such company exists');
+      throw new APIError('No such company exists', 404);
     }
 
     return result.rows[0];
@@ -100,6 +108,10 @@ class Job {
 
   // DELETE job by id
   static async deleteJobById(id) {
+    if (typeof id !== 'number') {
+      throw new APIError('Invalid id for job', 404);
+    }
+
     const result = await db.query(
       `DELETE FROM jobs
       WHERE id = $1
@@ -108,7 +120,7 @@ class Job {
     );
 
     if (result.rows.length === 0) {
-      throw new Error('No such company exists');
+      throw new APIError('No such company exists', 404);
     }
   }
 }
